requirements.txt
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudevents==0.3.0
Flask==1.1.2
functions-framework==2.0.0
google-api-core==1.22.2
google-auth==1.21.2
google-cloud-core==1.4.1
google-cloud-error-reporting==1.0.0
google-cloud-logging==1.15.1
google-cloud-pubsub==2.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
libcst==0.3.10
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
pathtools==0.1.2
proto-plus==1.9.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
watchdog==0.10.3
Werkzeug==1.0.1
wrapt==1.12.1
